Connectivity check before querying Foreign Security Principal
When the script starts, connectivity to all trusted domains is checked. When querying for Foreign Security Principals in these trusted domains, they might be located in a subdomain.

Due to the specific nature of Foreign Security Principals, the scripts need to perform a GC query against this subdomain. As this subdomain was not known before, the check makes sense.