Don't lock DLL file on server
Import-Module locks the DLL file on the server. Remove-Module can't remove this lock, as the underlying .Net function do not allow an actual unload of a DLL until the current PowerShell session is closed.

In this scenario, the DLL file is locked on the file server until all script instances are closed. Updating the DLL on the server would be complicated.

To work around this, the DLL is copied to the user's temp directory and loaded from there. As the DLL file is officially signed by Microsoft, this works fine in combination with AppLocker.