New: The configuration file is no longer a plain text file, but a full PowerShell script. This allows for complex replacement variable handling (complex string transformations, retrieving information from web services and databases, etc.). Attention: The configuration file is executed as part of Set-OutlookSignatures.ps1 and is not checked for any harmful content. Please only allow qualified technicians write access to this file, only use it to to define replacement variables and test it thoroughly. Attention: This change breaks compatibility with earlier versions. Custom configuration files have to be checked for PowerShell syntax and should have their extension changed from .txt to .ps1.
New: Simulation mode
Simulation mode is enabled when the parameter SimulatedUser is passed to the script. It answers the question "What will the signatures look like for user A, when Outlook is configured for the mailboxes X, Y and Z?".
In simulation mode, Outlook registry entries are not considered and nothing is changed in Outlook and Outlook web. The template files are handled just as during a real script run, but only saved to the folder passed by the parameters AdditionalSignaturePath and AdditionalSignaturePath folder.
Simulation mode is useful for content creators and admins, as it allows to simulate the behavior of the script and to inspect the resulting signature files before going live.
SimulationUser is a mandatory parameter for simulation mode. This value replaces the currently logged-on user.
SimulationMailboxes is optional for simulation mode, although highly recommended. It is a comma separated list of strings replacing the list of mailboxes otherwise gathered from the registry.
Active Directory data for both parameters is searched using Active Directory Ambigous Name Resolution (ANR), so you can pass very different values to finde the desired object (mail address, logon name, display name, etc.). Please see https://social.technet.microsoft.com/wiki/contents/articles/22653.active-directory-ambiguous-name-resolution.aspx for details about ANR.
Attention: Use values that are unique in an Active Directoy forest, not just in a domain. The script queries against the Global Catalog and always works with the first result returned only (even if there are additional results). For example, the logon name (sAMAccountName) must be unique within an Active Directory domain, but each domain in an Active Directory forest can have one account with this logon name - the results are returned in random order, the script always chooses the first result.

New: Major script steps now show a timestamp
ISO 8601 is used, with an @-sign left and right of the string.
Example: "@2021-07-21T10:31:55+02:00@" stands for the local date and time of the 21st of July 2021 at 10:31:55 AM, the local time differs +02:00 hours from UTC (Universal Time Coordinated).

New: Readme for "Simulation mode" and FAQ "How can I log the script output?"

Fixed: Templates with multiple groups or multiple mail addresses were not applied correctly and led to redundant Active Directory queries

Changed: When enumerating and categoring templates, category defining information is logged (assigned mail address, group name including SID)

Changed: Advanced primary mailbox detection and sorting
If the currently logged-on user has the AD attribute mail set, this mailbox receives signatures first.
If the mail attribute is not set, the script checks if exactly one of the mailboxes in the Outlook profiles has set the user as "Exchange external master account" and applies signatures to this mailbox first.
Mailboxes in the default outlook profile receive signature right after the primary mailbox, in the order they are stored in the registry (which is basically based on the time a mailbox has been added to a profile, oldest mailbox first, and totally independent from the sort order in Outlook).
Mailboxes from all other profiles receive signatures in the last step, in the order the are stored in the registry (which is basically based on the time a mailbox has been added to a profile, oldest mailbox first, and totally independent from the sort order in Outlook).

Changed: Easier readable output
Script start and end times are shown
Logical grouping of tasks (Outlook tasks first, basic AD tasks second, template enumeration third, then signature handling tasks)
Vertical whitespace between main tasks makes output easier consumable visually